added algorithm for decimal to roman conversion
